refactor(camera): meet every surface through one entry point, and name the rule
The Auto camera reaches the player through four generic files: the cfg reader
and writer, the console's cvar table, the console's command table, and the main
loop. Each of them named camera variables directly, so a new camera setting cost
an edit in two files and a new camera command an edit in a third.

Give the module an entry point per surface instead. FollowCam_RegisterCvars and
FollowCam_RegisterCommands join the FollowCam_ReadConfig / FollowCam_WriteConfig
pair, and cmd_camtrace moves in with them. Neither CONFIG_FILE.CPP nor
CONSOLE_CMD.CPP now names an Auto-camera variable; adding a camera setting is one
edit, in FOLLOWCAM.CPP.

Two cases deliberately stay where they are, because their state is not the Auto
camera's. cam_glide reads FollowCamOrbitGlide, which the manual-orbit
follow-through in EXTFUNC.CPP consumes. camnudge writes DbgCamNudge*, which the
same path consumes. Both stay on the bus until that path has a module.

CODESTYLE.md gains the general form the three extractions are instances of: a
feature is specific and owns state, a surface is generic and owns a calling
convention, and the surface must not name the feature's variables. CONTROL.CPP
is cited as the older and better precedent, since it already keeps its per-run
overrides in its own file and exposes Control_HasX / Control_GetX for consumers
to ask. The composability that falls out is already load-bearing: the camera
fixtures orbit the camera through CLI into control into console, with no
camera-specific test hook anywhere.

Registration order is not observable, so grouping the registrations changed
nothing a player sees: varlist and cmdlist sort by name.

Verified against the previous commit: varlist and cmdlist identical across all
99 lines, the written lba2.cfg byte-identical, --dump-state identical, and a
camtrace run identical for all 77 frames. 150/150 in the container; the camera
fixtures pass.
