+++ AGENTS.md
@@ -104,7 +104,7 @@ Apply these behavior rules on every non-trivial task:
 
 Style rules and rationale (language dialect, indentation, types, C++98, preservation) live in [CODESTYLE.md](CODESTYLE.md) — read it before writing engine code. The agent-critical invariant: **ported original code stays C-style to mirror the ASM; new infrastructure with no ASM original may use conservative C++98 features.** Never modernize game code (see the Never list). When unsure, match the file you are editing.
 
-**Adding a new subsystem: give it its own `.CPP`, and let it own its globals.** Do not grow an original file with code that never existed in 1997, and do not declare a new module's state in `C_EXTERN.H` / `GLOBAL.CPP`. The header declares exactly what the `.CPP` defines, so callers state their dependency in their include list. In a cross-cutting file (cfg, saves, state dumps) that rule inverts: it must not name another module's settings, so the module exposes a `Foo_ReadConfig` / `Foo_WriteConfig` pair and the cross-cutting file calls it. Full rule and rationale in [CODESTYLE.md "Where new code goes"](CODESTYLE.md#where-new-code-goes); [SOURCES/FOLLOWCAM.CPP](SOURCES/FOLLOWCAM.CPP) is the worked example. This matters because the dialect rule above is undecidable inside a file that is half original and half new.
+**Adding a new subsystem: give it its own `.CPP`, and let it own its globals.** Do not grow an original file with code that never existed in 1997, and do not declare a new module's state in `C_EXTERN.H` / `GLOBAL.CPP`. The header declares exactly what the `.CPP` defines, so callers state their dependency in their include list. In a *surface* (cfg, console, CLI, saves, state dumps: generic files that serve every feature) that rule inverts. A surface must not name a feature's variables; it offers a hook and the feature fills it in, as `Foo_ReadConfig` / `Foo_WriteConfig` / `Foo_RegisterCvars`. [SOURCES/CONTROL.CPP](SOURCES/CONTROL.CPP) is the same rule from the other side: it owns the per-run overrides and exposes `Control_HasX` / `Control_GetX` for consumers to ask. Full rule and rationale in [CODESTYLE.md "Where new code goes"](CODESTYLE.md#where-new-code-goes); [SOURCES/FOLLOWCAM.CPP](SOURCES/FOLLOWCAM.CPP) is the worked example. This matters because the dialect rule above is undecidable inside a file that is half original and half new.
 
 **Proving a refactor that should not change behaviour.** Relocation work is worth verifying by comparison, not by argument, and the engine's traces make that cheap: build the base commit and the change, run both under a pinned clock and identical input, and diff the trace. For camera work that is `--fixed-dt 16 --exec "camtrace 1" --exec-at <n> "camnudge ..." --tick <n> --exit`, diffed on the `[cam]` lines; other subsystems have their own (`objtrace`, `lifetrace`, `--dump-state`, `--projrec`). A byte-identical trace is a stronger result than any test asserting the same thing, and it costs one rebuild.
