+++ CODESTYLE.md
@@ -53,6 +53,32 @@ Two things to expect when doing this:
 - **A serialised format's ordering can outrank the rule.** `lba2.cfg` is rewritten in key order, so a module can only take a contiguous run of keys without reordering every player's file. Where a key sits apart from its module's block, leave it and say why, as `FollowCamera` is left in FOLLOWCAM.H. Cosmetic churn in a user's file is a worse outcome than an imperfect boundary.
 - **A module's private state usually has one leak.** Both extractions found exactly one: the main loop assigning the camera's smoothing state, and boot assigning the cfg reader's `StoredLanguage`. Give it a named entry point rather than widening the header to expose the variable. The name is the documentation for a coupling that was previously invisible.
 
+### Features and surfaces
+
+The two rules above are the same rule seen from either end, and naming the two kinds of file says when to apply which.
+
+A **feature** is specific: the Auto camera, the save system, resolution switching. A **surface** is generic and serves every feature: the cfg reader and writer, the console, the CLI flag table and the control harness, the options menu. A feature has state and behaviour; a surface has a calling convention and no opinion about who uses it.
+
+That asymmetry fixes the dependency direction. **A surface must not name a feature's variables; it offers a registration or serialisation hook, and the feature fills it in.** Otherwise every new setting costs an edit in each generic file, and those edits are what turned the entry point into a 4000-line file in the first place.
+
+The Auto camera is the worked example because it now meets every surface this way, each through one entry point in [SOURCES/FOLLOWCAM.CPP](SOURCES/FOLLOWCAM.CPP):
+
+| Surface | Entry point |
+|---|---|
+| lba2.cfg | `FollowCam_ReadConfig` / `FollowCam_WriteConfig` |
+| Console cvars | `FollowCam_RegisterCvars` |
+| Console commands | `FollowCam_RegisterCommands` |
+| Main loop | `UpdateFollowCameraExt`, `FollowCamTrace`, `FollowCamResetToView`, `FollowCamSyncTarget` |
+
+Neither [SOURCES/CONFIG_FILE.CPP](SOURCES/CONFIG_FILE.CPP) nor `CONSOLE/CONSOLE_CMD.CPP` names an Auto-camera variable. Adding a camera setting is one edit, in the module.
+
+Two things worth knowing before applying this:
+
+- **Some surfaces already work this way, and are the better precedent.** [SOURCES/CLI_ARGS.CPP](SOURCES/CLI_ARGS.CPP) names no engine state at all; it is a flag table for `--help` and unknown-flag rejection. [SOURCES/CONTROL.CPP](SOURCES/CONTROL.CPP) keeps every per-run override in its own file-scope state and exposes `Control_HasX` / `Control_GetX`, which is this rule from the other side: the owner offers the accessor, so `CONFIG_FILE.CPP` asks rather than reaches.
+- **The payoff is composability, and it is already load-bearing.** Because the console is a surface rather than a set of special cases, the camera can be driven through it without a camera-specific test hook: the fixtures in `tests/automation` orbit the camera with `--exec "camnudge ..."`, which is CLI into control into console into the feature. A feature that meets its surfaces properly becomes testable for free.
+
+The rule is about ownership of storage, not about forbidding reads. `GAMEMENU.CPP` reading `FollowCamera` through FOLLOWCAM.H to choose a menu label is a plain use of a public setting and needs no hook.
+
 ## Layout and naming
 
 - **Indentation:** 4 spaces in C/C++. Tabs are preserved in ASM. The original used tabs; there is an ongoing migration to 4 spaces, so new contributions use 4 spaces.
